Added jackpot command. Simplified Profile system.
